Created a comment service and transfered functions. also got the delete comment functionally
