+++ src/components/services/postService.js
@@ -14,23 +14,6 @@ export const getAllFormats = () => {
   return fetch("http://localhost:8000/formats").then((res) => res.json());
 };
 
-export const getAllComments = () => {
-  return fetch(`http://localhost:8000/comments/?_expand=user`).then((res) =>
-    res.json()
-  );
-};
-
-export const submitNewComment = (newComment) => {
-  const postOptions = {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(newComment),
-  };
-  return fetch("http://localhost:8000/comments", postOptions);
-};
-
 export const createNewPost = (newPost) => {
   const postOptions = {
     method: "POST",
